+++ src/solution.c
@@ -157,7 +157,7 @@ void obs_callback(u16 sender_id, u8 len, u8 msg[], void* context)
   /* Calculate packet latency */
   if (time_quality >= TIME_COARSE) {
     gps_time_t now = get_current_time();
-    s32 latency_ms = (s32) ((now.tow - t.tow) * 1000.0);
+    float latency_ms = (float) ((now.tow - t.tow) * 1000.0);
 
     log_obs_latency(latency_ms);
   }
